HttpClientResponseException.getMessage() returns null when the body is not empty
When the response's body is not empty but it cannot be converted to
JsonError the method getMessage() get a JsonError with a null message.
So instead of printing the exception message it prints null.

This commit fixes that by checking if the JsonError's message is null
and in that case it prints the exception's message instead.