Preserve MDC original context after instrumented operation completion (#2316)
Currently, MDC instrumenter always clear the MDC context at the end of
the instrumented execution cycle.  When an Async operation is run as blocking,
the entire execution happens in same thread as the calling thread. At the end of the
execution of the asynchronous operation, the current thread MDC will be cleared, which
can lead to unexpected loss of context.

This change proposes that during an instrumented execution, the MDC
context is only overloaded and at the end of the execution the previous
state is restored.