Introduce `@ContextConfigurer` annotation (#6386)
* Introduce `@ContextConfigurer` annotation

This new annotation allows application context builder customizers
automatically. A class implementing `ApplicationContextCustomizer`
which is _also_ annotated with `@ContextConfigurer` will automatically
be registered as a service, and loaded when an application builder
is created.

There are several advantages of doing this:

- the "things" which are responsible for configuring the application
context become discoverable to other tools, in particular the new
Micronaut AOT module (so it can execute them on its own)
- the configuration blocks can be shared, or even published, and
would be automatically applied to all contexts created as long as
the service is found on classpath

Hence, a recommended pattern to configure an application would be
to rewrite this:

```java
class Application {
    public static void main(String... args) {
          Micronaut.builder(args)
             .mainClass(Application.class)
             .deduceEnvironment(false)
             .start();
    }
}
```

You would now write:

```
@MicronautApplication
public class Application implements ApplicationContextCustomizer {
     @Override
     public void configure(ApplicationContextBuilder builder) {
         builder.deduceEnvironment(false);
     }

     public static void main(String... args) {
         Micronaut.run(args, Application.class);
     }
}
```

While being longer to write, this patterns clearly
separates the configuration phase from the execution
phase.

If the application class is also used as a bean, a
recommended pattern would be to use an inner class:

```
@Singleton
public class Application {
    @ContextConfigurer
    public static class Configurer implements ApplicationContextCustomizer {
        @Override
        public void configure(ApplicationContextBuilder builder) { ... }
    }

    ...
}
```

There are 2 annotations which currently have the same semantics:

- `@ContextConfigurer` is the main one, used to annotate any
application context builder configurer
- `@MicronautApplication`, which is used when annotating an
application class. While it's also possible to annotate such
a class with `@ContextConfigurer`, there's a semantics difference.
In the future this annotation may be used to generate more code,
in particular entry points.

* Fix review feedback

* Fix compatibility with Java 8 compiler

* Mark APIs as experimental

* Remove `@MicronautApplication`

Since for now it doesn't bring any advantage over `@ContextConfigurer`.

* Verify that classes annotated with @ContextConfigurer have no ctor

Classes annotated with `@ContextConfigurer` are good old Java services,
which are not supposed to have constructors with args. This commit makes
sure that it is the case by failing compilation if such constructors are
found.

* Update error message for clarity