Fix binding mixed attributes to body parameters (#6764)
`NettyHttpRequest.buildBody` distinguishes between requests with "received data" and "received content", where the former only applies when `AbstractHttpData` instances are present. `AbstractHttpData` is a base class of both `DiskAttribute` and `MemoryAttribute`, but not of `MixedAttribute`. This makes `buildBody` use the "content" path for mixed requests.

This change adds an exception for `MixedAttribute` so that it is also covered by the "data" path.

Fixes #6705