Release unwrapped body in NettyHttpRequest (#6891)
* Release unwrapped body in NettyHttpRequest
This is an attempt at a fix for the flaky FuzzyInputSpec. Can't reproduce that locally, but this code is obviously buggy, so this might help.

`body` is a memoized supplier, but `NettyHttpRequest.release` checks whether it's a reference counted object. The supplier is never reference counted, but the object it supplies might be. This patch stores the unwrapped value as a separate field so that it can be released properly.

The test for #6712 still passes, so this doesn't seem to regress the issue that the patch that caused this leak fixed.

* fix setBody

* Release buffers in tests

Co-authored-by: jameskleeh <james.kleeh@gmail.com>