Fix response leak when request is already done (#6935)
If HandlerPublisher is in DONE state when the RoutingInBoundHandler would subscribe to it in writeFinalNettyResponse, neither onComplete nor onError would be called, and the response would never be written. This manifested in a buffer leak in fuzzing, when the response contained a ByteBuf body.

The change to RoutingInBoundHandler is unrelated, it's simply about discarding the wrapper micronaut ByteBuffer when it's not used anymore. It's a no-op right now, but we don't specify anywhere that ByteBuffer just delegates the ref counting to the netty buffer, so imo this is a sensible change.