fix: allow deserializing _embedded values of HATEOAS resources (#6850)
Adds a `GenericResource` class that acts as a deserialization target for the embedded resources. We can't use "real" resource classes because there's no way to tell which implementation is the right one.

It's possible that we could use jackson "deduction" subtype detection to support common implementations like `JsonError`, to give the right resource type in at least some cases. However this jackson feature isn't supported by serde, so I've not looked at this.

Fixes #2863

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>