Rework BOM as core BOM (#8299)
* Remove dependencies which are not directly used

This commit removes from the core BOM all managed dependencies which
are not directly used in build scripts. This basically means that
those dependencies are only present so that precisely we can have them
in the BOM. Now that the BOM is extracted in its own project, they are
redundant.

* Remove managed libraries which are not used

This commit moves a number of libraries from the managed scope to unmanaged.
The only dependencies which are managed are now the ones which are in api
or implementation scope.

Those dependencies were previously in the BOM, but won't be directly managed
by core. This means a couple of things:

1. the new BOM will have to duplicate those dependencies, in order to manage
them
2. the dependencies will have to be copied in the new BOM (because they are
not present in the core BOM anymore, they won't be inlined like the other
managed dependencies)

* Rename `bom` -> `core-bom`

This is a breaking change: it renames the `bom` project to `core-bom`.
The new BOM will live in a separate project (`micronaut-bom`) under
different GAV coordinates.

* Reintroduce missing BOMs

* Fix Groovy BOM coordinates

* Update groovy-test coordinates

* Restore entries for inject-xxx in core BOM

* Revert "Update groovy-test coordinates"

This reverts commit b376b79d69aa89da3d6b234d465c13746e8aa55b.