Detect deadlocks when using BlockingHttpClient on the event loop (#8354)
Keep track of the thread that called BlockingHttpClient.exchange, and throw an exception when the client attempts to use the same thread for the connection. Before this patch, the blocking call would time out, now it will fail with a specific exception.

The event loop group usually has many threads, and often a blocking call will wait on *another* thread of that event loop group by chance. Such uses currently only lead to sporadic read timeouts at the moment. This implementation is conservative: It will not fail when the same event loop group is used, it will only fail with the exact same thread. So the sporadic timeouts will be replaced by sporadic deadlock exceptions.

Relates to #8198

Co-authored-by: Tim Yates <tim.yates@gmail.com>