Cleanup GraalVM metadata and used Metadata Repository (#8821)
With the GraalVM Metadata Repository now active we no longer need metadata for Netty and Logback so this PR removes that metadata. In addition this PR removes the use of all GraalVM internal APIs from core and upgrades to the latest version of SVM.

Finally, this PR deprecates AutomaticFeatureUtils because in future it should not be necessary to use this class with metadata living the repository.