Enable tcnative even when ALPN is not configured (#9397)
This PR enables tcnative support when available. Before this patch it would only be turned on if ALPN was necessary.

I also switched from OPENSSL to OPENSSL_REFCNT, which is more performant, and moved to the openssl-specific cached key manager factories from netty. To test this, I added tcnative to the http-server-netty tests, which also has leak detection.