By default, add JSON additional Types (#9406)
We have several modules, e.g. problem+json, which require a JSON additional type.

This PR creates a constant in the JsonMediaTypeCodec with the additional types which I think is the correct way of defining those built in additional types and it adds a test which verifies the NettyJsonHandler and JsonMessageHandler produces and consumes annotation match that constant.