Replace HttpStreamsHandler for netty HTTP client (#9483)
Just like #9128 did for the server, this PR replaces the HttpStreamsHandler and HandlerPublisher for the client. Instead, incoming HttpRequest and HttpContent objects are handled directly and transformed to the needed types.

This allows full removal of the streams handler. Some of the functionality is reimplemented in ReactiveClientReader and ReactiveClientWriter, but these are much simpler classes and only need to be used for streaming requests / responses.

This implementation is not quite as optimized as PipeliningServerHandler, there's no read coalescing for example, and it uses multiple handlers where a more complex implementation could do with one.

---------

Co-authored-by: Milenko Supic <supic82@gmail.com>