Fix status route resolution (#9851)
This patch fixes an error with a >400 status returned by a controller.

I moved a bunch of API to int-based status handling. However, this PR does not go all the way: We may want to add int-based parameters to `@Status` and `@Error`. But that is more work especially to test so I left it out of this PR.

Fixes #9824