Fix file uploads beyond MMAP_SEGMENT_SIZE (#10793)
* Fix file uploads beyond MMAP_SEGMENT_SIZE
The code path was not covered by tests and was broken.

I also replaced the ReentrantLock by a NonReentrantLock that is based on a Semaphore. This is because there are some semantic checks in Chunk that might not work with a reentrant lock (e.g. repeated deallocate calls could lead to ref count errors). But this is unrelated to the bug.

Fixes #10666

* Update http-server-netty/src/main/java/io/micronaut/http/server/netty/NonReentrantLock.java

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>

* make badssl more resilient to timeouts

* make badssl more resilient to timeouts

---------

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>