Fix build with JCenter removal (#10990)
JCenter is now redirecting to Maven Central, so it doesn't
serve artifacts which were only in JCenter anymore. Of course,
our build had such an artifact. It was brought through the
"webdrivers" plugin, which depends on a grolifant version which
is not on Maven Central.

Unfortunately, upgrading that plugin is not straightforward.
While there is an alternative, there is actually no version
which is compatible with Gradle 8.7:

https://github.com/erdi/webdriver-binaries-gradle-plugin/issues/43

At the same time, the `test-geb` module was actually not
tested on CI. The only way to run the tests was locally, and the
instructions to do so were wrong, so it's unlikely that it gave
any useful signal anywhere recently.

As such, this commit removes the `test-suite-geb` module, at
least until we have a proper replacement.

Fixes #10988