Skip classes from introspection generation that are postponed (#11155)
Currently if a class is postponed the introspection is still generated. This means that in the next round the introspection is recreated and an exception is thrown because there was an attempt to rewrite the some class.

Without this change it is impossible to use @Wither and @Introspected together.

Unclear how to create a test for it here. Will have to upgrade SourceGen and add a test there.

---------

Co-authored-by: Denis Stepanov <denis.s.stepanov@oracle.com>