Improve event loop locality of client requests (#11300)
If a request is made on an event loop that is part of the same event loop group that the client is configured to use, prefer making the connection on that same event loop. This keeps all the relevant processing on the same thread and may improve performance.If a request is made on an event loop that is part of the same event loop group that the client is configured to use, prefer making the connection on that same event loop. This keeps all the relevant processing on the same thread and may improve performance.