Don't clear metadata already cleared by type element visitor (#11808)
Currently if a type element visitor throws element postponed and then in the next round mutates the metadata then this metadata is cleared by the bean definition inject visitor causing it to be missing. This PR alters the flow such that the type element visitor re-initializes the metadata marking it as cleared and thus the bean definition inject visitor doesn't also clear it.

Fixes #11791