Avoid double completion in DefaultHttpClient (#11894)
There are a few cases in DefaultHttpClient where a connection error could lead to two exceptions, with both sent to the DelayedExecutionFlow. This patch adds an atomic tryCompleteExceptionally method to avoid this.

The new tryComplete methods are thread-safe, but the old complete methods remain unsafe to avoid performance regressions. I don't think the atomicity is even really necessary for DefaultHttpClient, but it can't hurt.