Use ReadBuffer to reduce code duplication (#12024)
* Introduce ReadBuffer API
This PR adds a new ReadBuffer API that aims to replace the current ByteBuffer.

The main advantage is a clear ownership system based on Closeable instead of the optional ReferenceCounted. This simplifies user code, and gets rid of potentially slow `instanceof` checks. It's also necessary for netty 5 compatibility.

As the name suggests, a `ReadBuffer` is read-only. The ByteBuffer write methods are rarely used by micronaut projects, so they only add unnecessary complexity for the non-netty implementation. The few write use cases will be covered by the ByteBodyFactory buffer methods.

In addition to the new API, I also moved `AvailableByteBody` to use `ReadBuffers`. In the future, we will only need one implementation of AvailableByteBody based on ReadBuffer, instead of the array and netty based ones that exist right now.

My plan is to use ReadBuffers to also merge some of the more complex ByteBody implementations, or make them more efficient. For example, ReactiveByteBufferByteBody could use ReadBuffers instead of nio ByteBuffers, and BaseSharedBuffer could be simplified using ReadBuffers instead of having to accommodate both netty and nio buffers.

* make InternalByteBody an abstract class

* docs

* remove unreachable null check

* CR

* rename readSome to split

* add ReadBuffer.transferTo

* add ReadBuffer.toString

* add ReadBufferFactory.compose

* Use ReadBuffer to reduce code duplication
Lots of ByteBody code is duplicated for netty and non-netty. The common ReadBuffer API makes a refactor to merge much of that code possible.

- The reactive implementation of AbstractBodyAdapter is merged into the base class, which becomes non-abstract. The netty implementation retains some small thread safety additions.
- BaseSharedBuffer is refactored to work with ReadBuffers directly. Much of the subclass code is merged into the base class. Subclasses are retained due to different approaches to thread safety.
- BaseSharedBuffer.AsFlux subclasses are removed.
- A common class BaseStreamingByteBody is introduced that is based on BaseSharedBuffer. Some of the code from the subclasses is moved there.
- BufferConsumer gets an add(ReadBuffer) method. Subinterfaces with add(ByteBuf) or add(ByteBuffer) become obsolete and are removed.
- ByteBodyFactory gets adapt methods for converting publishers to ByteBodys, previously done through NettyBodyAdapter (mostly emptied of code) and ByteBufferBodyAdapter (deprecated).
- ConcatenatingSubscriber deals with ReadBuffers directly. ByteBuffer subclasses are removed. Netty subclass remains for concurrency control. Json-specific subclasses are replaced by a Separators record.
- PublisherAsBlocking is ungenerified and only deals with ReadBuffer now. Subclasses with specialized resource management are removed.
- PublisherAsStream is moved to ReadBuffers and is moved to micronaut-http.
- micronaut-http-client-jdk ByteBodySubscriber is mostly cleared and uses ByteBodyFactory.adapt to transform the reactive stream to a ByteBody.
- NettyByteBody is removed because it didn't do very much anymore, so that StreamingNettyByteBody can extend BaseStreamingByteBody instead.

* test change

* trigger build

* fix merge