Clarify documentation about underscores in properties (#12035)
Due to the way micronaut translates snake case environment variables to kebab case properties, it is not possible to target a property with underscores using environment variables.

Solutions could exist for this by normalizing `_` to `.` or `-` in regular property processing. However, until then, simply documenting this limitation explicitly will be helpful.