Add explicit event loop transport configuration (#11987)
prefer-native-transport only sets a preference. If the native transport fails to load for any reason (e.g. wrong architecture), the event loop silently falls back to nio.

This PR adds a new property to set an explicit preference list of transports to use. If the user wants to be sure they aren't using nio by accident, they can simply remove the nio entry from the list and there will be an exception.