Introduce a Micronaut module descriptor (#12192)
* Introduce a Micronaut module descriptor

This commit introduces a facility to list the Micronaut modules
which are found at runtime. This will be useful diagnostics
information, which can for example be used at runtime by the
Micronaut Control Panel.

It is implemented 2-fold. First of all, an independent module
named `micronaut-module-info`, standalone without any dependencies,
defines the base interface and abstract class that can be used
for such descriptors. The goal is to load the descriptors using
service loading, independently of the framework. The reason we
do this is because this module can then be used by the build
plugins to either:
   - generate an implementation of a descriptor, which will
then be loaded by Micronaut
   - statically reason about modules, for example for linting
at build time.

The second module, `micronaut-module-info-runtime`, maps these
services into Micronaut beans. A singleton, `MicronautRuntimeModules`
can be used to fetch information about modules, which are
then resolved as a hierarchy of modules.

For now, the implementation is limited, provides minimal information
such a name for the module, its version, the Maven coordinates of
the module and some tags. Eventually, we may enrich this with more
information such as the dependencies of the module, or compatibility
with other modules. For example this could become possible to express
runtime requirements such as "module A requires at least one implementation
of B at runtime, but not more than 1".

Or, it could become possible for a module to explain which test resources
modules it requires, or even supply a list of MCP servers that could be
used for coding with AI...

* Disable ossIndexAudit on module-info module

* Address review feedback

* test: isolate ConstantPropertySourceSpec in its own module (#12195)

test ConstantPropertySourceSpec passes when run on isolation. It fails when run in a suite. This PR moves the test to its own module. Not ideal but at least we remove the flakiness of the test:

https://ge.micronaut.io/scans/tests?search.relativeStartTime=P90D&search.timeZoneId=Europe%2FMadrid&tests.container=io.micronaut.context.env.ConstantPropertySourceSpec&tests.help=tags

* test: ProxyBackpressureTest run localy reduced cvssource (#12196)

Tests with:

```
        "true,2,/large",
        "true,2,/proxy",
``

Fail locally:

https://ge.micronaut.io/scans/tests?search.timeZoneId=Europe%2FMadrid&tests.container=io.micronaut.http.client.ProxyBackpressureTest&tests.help=tags

While they are flaky in CI:

https://ge.micronaut.io/scans/tests?search.tags=CI&search.timeZoneId=Europe%2FMadrid&tests.container=io.micronaut.http.client.ProxyBackpressureTest&tests.help=tags

This pull-request modifies the test so that locally only a reduce CSV source is run

* test: ignore LoomCarrierSpec for MacOS (#12197)

* test: ignore LoomCarrierSpec for MacOS

LoomCarrierSpec fails on MacOS passes for Linux

https://ge.micronaut.io/scans/tests?search.startTimeMax=1763506799999&search.startTimeMin=1762815600000&search.tags=not:CI&search.timeZoneId=Europe%2FMadrid&tests.container=io.micronaut.http.server.netty.LoomCarrierSpec&tests.test=sticky%20on%20poller%20thread&tests.unstableOnly=false

* ignore if Java 25

* Update common files (#12178)

* build: disable htmlSanityCheck (#12198)

see: https://github.com/micronaut-projects/micronaut-core/issues/12181

---------

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>
Co-authored-by: micronaut-build <65172877+micronaut-build@users.noreply.github.com>